casi todo

+++ src/Controller/UsersController.php
@@ -10,6 +10,19 @@ namespace App\Controller;
  */
 class UsersController extends AppController
 {
+
+    public function initialize(): void
+    {
+        parent::initialize();
+        $this->loadComponent('Authentication.Authentication');
+    }
+
+    public function beforeFilter(\Cake\Event\EventInterface $event)
+    {
+        parent::beforeFilter($event);
+        $this->Authentication->addUnauthenticatedActions(['login', 'logout']);
+    }
+
     /**
      * Index method
      *
@@ -46,7 +59,14 @@ class UsersController extends AppController
     {
         $user = $this->Users->newEmptyEntity();
         if ($this->request->is('post')) {
-            $user = $this->Users->patchEntity($user, $this->request->getData());
+            $data = $this->request->getData();
+
+            if (!empty($data['password'])) {
+                $data['password'] = password_hash($data['password'], PASSWORD_DEFAULT);
+            }
+
+            $user = $this->Users->patchEntity($user, $data);
+
             if ($this->Users->save($user)) {
                 $this->Flash->success(__('The user has been saved.'));
 
@@ -68,8 +88,17 @@ class UsersController extends AppController
     public function edit($id = null)
     {
         $user = $this->Users->get($id, contain: []);
-        if ($this->request->is(['patch', 'post', 'put'])) {
-            $user = $this->Users->patchEntity($user, $this->request->getData());
+        if ($this->request->is(['patch', 'put'])) {
+            $data = $this->request->getData();
+
+            if (!empty($data['password'])) {
+                $data['password'] = password_hash($data['password'], PASSWORD_DEFAULT);
+            } else {
+                unset($data['password']);
+            }
+
+            $user = $this->Users->patchEntity($user, $data);
+
             if ($this->Users->save($user)) {
                 $this->Flash->success(__('The user has been saved.'));
 
@@ -90,14 +119,48 @@ class UsersController extends AppController
      */
     public function delete($id = null)
     {
-        $this->request->allowMethod(['post', 'delete']);
+        $this->request->allowMethod(['delete']);
+        $currentUserId = $this->request->getSession()->read('Auth.User.id');
+
+        $currentUser = $this->Users->get($currentUserId, [
+            'contain' => ['Profiles']
+        ]);
+
+        if ($currentUser->profile->role !== 'admin') {
+            $this->Flash->error('You are not authorized to delete users.');
+            return $this->redirect(['action' => 'index']);
+        }
+
         $user = $this->Users->get($id);
         if ($this->Users->delete($user)) {
-            $this->Flash->success(__('The user has been deleted.'));
+            $this->Flash->success('User deleted.');
         } else {
-            $this->Flash->error(__('The user could not be deleted. Please, try again.'));
+            $this->Flash->error('User could not be deleted.');
         }
 
         return $this->redirect(['action' => 'index']);
     }
+
+    public function login()
+    {
+        if ($this->request->is('post')) {
+            $user = $this->Users->find()
+                ->where(['email' => $this->request->getData('email')])
+                ->first();
+
+            if ($user && password_verify($this->request->getData('password'), $user->password)) {
+                $this->request->getSession()->write('Auth.User', $user);
+                return $this->redirect(['controller' => 'Users', 'action' => 'index']);
+            } else {
+                $this->Flash->error('Invalid email or password');
+            }
+        }
+    }
+
+    public function logout()
+    {
+        $this->Authentication->logout();
+        return $this->redirect(['controller' => 'Users', 'action' => 'login']);
+    }
+
 }
